% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPhecodeInfo.R
\name{addPhecodeInfo}
\alias{addPhecodeInfo}
\title{Add Phecode Information}
\usage{
addPhecodeInfo(d, pheinfo_table = NULL)
}
\arguments{
\item{d}{A data frame with a \code{phecode} column}

\item{pheinfo_table}{The phecode info table. Defaults to extended_pheinfo.}
}
\value{
The input data frame with added columns: description, group,
  groupnum, color
}
\description{
Adds phecode descriptions, group names, group numbers, and colors to a
data frame that contains a phecode column.
}
