% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_matcher.R
\docType{data}
\name{ANATOMICAL_MAP}
\alias{ANATOMICAL_MAP}
\title{Anatomical keyword to phecode range mapping}
\format{
An object of class \code{list} of length 84.
}
\usage{
ANATOMICAL_MAP
}
\description{
Anatomical keyword to phecode range mapping
}
\keyword{internal}
