% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPhenotypes.R
\name{apply_sex_restrictions}
\alias{apply_sex_restrictions}
\title{Apply Sex Restrictions to Phenotype Table}
\usage{
apply_sex_restrictions(phenotypes, id.sex, sex_restrict, id_col)
}
\arguments{
\item{phenotypes}{Wide-format phenotype data frame}

\item{id.sex}{Data frame with id and sex columns}

\item{sex_restrict}{Sex restriction table}

\item{id_col}{Name of the ID column}
}
\value{
Modified phenotype data frame
}
\description{
Apply Sex Restrictions to Phenotype Table
}
\keyword{internal}
