% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_matcher.R
\name{assign_procedure_phecode}
\alias{assign_procedure_phecode}
\title{Assign New Procedure Phecode}
\usage{
assign_procedure_phecode(opcs_code, chapter)
}
\arguments{
\item{opcs_code}{The OPCS code (e.g., "A01")}

\item{chapter}{The OPCS chapter letter}
}
\value{
Character string phecode
}
\description{
For OPCS codes that don't map to existing condition phecodes, assign a new
procedure phecode in the 1000+ range.
}
\keyword{internal}
