% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_opcs_phecode_map.R
\name{build_opcs3_phecode_map}
\alias{build_opcs3_phecode_map}
\title{Build OPCS-3 to Phecode Mapping}
\usage{
build_opcs3_phecode_map(opcs3_df, pheinfo, opcs4_map = NULL)
}
\arguments{
\item{opcs3_df}{Data frame from parse_opcs3()}

\item{pheinfo}{Data frame of existing phecode info}

\item{opcs4_map}{Existing OPCS-4 mapping (to chain OPCS-3 -> OPCS-4 where possible)}
}
\value{
List with map, audit, new_phecodes components
}
\description{
Same approach as OPCS-4 but for the older OPCS-3 code set.
}
