% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_extended_maps.R
\name{build_procedure_exclusions}
\alias{build_procedure_exclusions}
\title{Build Procedure Phecode Exclusions}
\usage{
build_procedure_exclusions(new_phecodes, group_nums)
}
\arguments{
\item{new_phecodes}{Data frame of new procedure phecodes}

\item{group_nums}{Named vector of group numbers}
}
\value{
Data frame with code, exclusion columns
}
\description{
Build Procedure Phecode Exclusions
}
\keyword{internal}
