% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_matcher.R
\name{chapter_to_phecode_base}
\alias{chapter_to_phecode_base}
\title{OPCS Chapter to Procedure Phecode Base Number}
\usage{
chapter_to_phecode_base(chapter)
}
\arguments{
\item{chapter}{Single character A-X}
}
\value{
Integer base phecode number
}
\description{
Maps OPCS-4 chapter letters to the base phecode number in the 1000+ range.
}
\keyword{internal}
