% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_read_v3.R
\name{classify_read_code}
\alias{classify_read_code}
\title{Classify Read Code Type}
\usage{
classify_read_code(codes)
}
\arguments{
\item{codes}{Character vector of Read codes}
}
\value{
Character vector of code types
}
\description{
Classify Read Code Type
}
\keyword{internal}
