% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_snomed.R
\name{classify_snomed_hierarchy}
\alias{classify_snomed_hierarchy}
\title{Classify SNOMED Semantic Tag into Hierarchy}
\usage{
classify_snomed_hierarchy(tags)
}
\arguments{
\item{tags}{Character vector of semantic tags}
}
\value{
Character vector of hierarchy classifications
}
\description{
SNOMED FSNs end with semantic tags like (disorder), (procedure),
(finding), (body structure), etc.
}
\keyword{internal}
