% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_extended_maps.R
\name{generate_procedure_colors}
\alias{generate_procedure_colors}
\title{Generate Distinct Colors for Procedure Groups}
\usage{
generate_procedure_colors(n)
}
\arguments{
\item{n}{Number of colors needed}
}
\value{
Character vector of hex color codes
}
\description{
Generate Distinct Colors for Procedure Groups
}
\keyword{internal}
