% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCodesToPhecodes.R
\name{get_package_data}
\alias{get_package_data}
\title{Get Package Data Object}
\usage{
get_package_data(name)
}
\arguments{
\item{name}{Name of the data object}
}
\value{
The data object
}
\description{
Loads a data object from the package's data directory.
}
\keyword{internal}
