% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPhecodeInfo.R
\name{mappingSummary}
\alias{mappingSummary}
\title{Mapping Coverage Summary}
\usage{
mappingSummary(vocabulary.map = NULL)
}
\arguments{
\item{vocabulary.map}{The mapping table. Defaults to extended_phecode_map.}
}
\value{
A data frame summarizing mapping coverage
}
\description{
Reports the number of codes and phecodes by vocabulary type.
}
