% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCodesToPhecodes.R
\name{normalize_vocab_id}
\alias{normalize_vocab_id}
\title{Normalize Vocabulary ID Strings}
\usage{
normalize_vocab_id(x)
}
\arguments{
\item{x}{Character vector of vocabulary IDs}
}
\value{
Character vector of normalized IDs
}
\description{
Converts various representations of vocabulary IDs to the canonical form.
}
\keyword{internal}
