% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_icd9.R
\name{parse_icd9_procedures}
\alias{parse_icd9_procedures}
\title{Parse ICD-9 Procedure Codes}
\usage{
parse_icd9_procedures(filepath)
}
\arguments{
\item{filepath}{Path to the ICD9 Operative Codes.RTF file}
}
\value{
A data frame with columns: code, description
}
\description{
Parses the ICD-9 Operative Codes RTF file into a data frame.
}
