% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_opcs4.R
\name{parse_opcs4}
\alias{parse_opcs4}
\title{Parse OPCS-4 Codes}
\usage{
parse_opcs4(filepath)
}
\arguments{
\item{filepath}{Path to the OPCS 4.rtf file}
}
\value{
A data frame with columns: code, description, node_id, parent_id,
  selectable, chapter
}
\description{
Parses the OPCS-4 RTF file into a clean data frame of procedure codes.
}
