% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_read_v2.R
\name{parse_read_v2_csv}
\alias{parse_read_v2_csv}
\title{Parse Read v2 from CSV file}
\usage{
parse_read_v2_csv(filepath, code_col, desc_col)
}
\description{
Parse Read v2 from CSV file
}
\keyword{internal}
