% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_read_v2.R
\name{parse_read_v2_trud}
\alias{parse_read_v2_trud}
\title{Parse Read v2 from NHS TRUD Release Directory}
\usage{
parse_read_v2_trud(dirpath, include_synonyms = FALSE)
}
\arguments{
\item{dirpath}{Path to the TRUD release directory}

\item{include_synonyms}{Include synonym terms}
}
\value{
Data frame with code and description
}
\description{
NHS TRUD releases contain multiple files. The key files are:
- Unified/V2/Corev2.all: Complete code list
- Or individual chapter files
}
\keyword{internal}
