% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_read_v2.R
\name{parse_read_v2_tsv}
\alias{parse_read_v2_tsv}
\title{Parse Read v2 from TSV file}
\usage{
parse_read_v2_tsv(filepath, code_col, desc_col)
}
\description{
Parse Read v2 from TSV file
}
\keyword{internal}
