% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_snomed.R
\name{parse_snomed_flat}
\alias{parse_snomed_flat}
\title{Parse SNOMED-CT from a Flat File}
\usage{
parse_snomed_flat(filepath)
}
\description{
For simplified exports or custom SNOMED code lists.
}
\keyword{internal}
