% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPhecodeInfo.R
\name{restrictPhecodesBySex}
\alias{restrictPhecodesBySex}
\title{Restrict Phecodes by Sex}
\usage{
restrictPhecodesBySex(phecodes, sex)
}
\arguments{
\item{phecodes}{Character vector of phecodes}

\item{sex}{Character. "M" for male, "F" for female.}
}
\value{
Filtered character vector of phecodes appropriate for the given sex
}
\description{
Applies sex-specific restrictions to a list of phecodes.
}
