% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_matcher.R
\name{score_substring_overlap}
\alias{score_substring_overlap}
\title{Score Substring Overlap Between Two Descriptions}
\usage{
score_substring_overlap(clinical_terms, phecode_desc)
}
\arguments{
\item{clinical_terms}{Cleaned OPCS clinical terms}

\item{phecode_desc}{Phecode description}
}
\value{
Numeric score between 0 and 1
}
\description{
Calculates a similarity score based on shared meaningful words.
}
\keyword{internal}
